requirements.txt
asgiref==3.3.1
attrs==20.3.0
awsebcli==3.19.3
bcrypt==3.2.0
blessed==1.17.12
botocore==1.19.63
cached-property==1.5.2
cement==2.8.2
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
colorama==0.4.3
cryptography==3.4.6
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-cors-middleware==1.5.0
django-heroku==0.3.1
django-ipware==3.0.2
docker==4.4.3
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
future==0.16.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.4.0
jmespath==0.10.0
jsonschema==3.2.0
paramiko==2.7.2
pathspec==0.5.9
psycopg2==2.7.7
pycparser==2.20
PyNaCl==1.4.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
PyYAML==5.3.1
requests==2.24.0
semantic-version==2.5.0
six==1.15.0
sqlparse==0.4.1
termcolor==1.1.0
texttable==1.6.3
typing-extensions==3.7.4.3
urllib3==1.25.11
wcwidth==0.1.9
websocket-client==0.57.0
whitenoise==5.2.0
zipp==3.4.0
